move bindCard to a new file
